Refactor hannk's compare_vs_tflite code to be mostly library (#5991)
* Refactor compare_vs_tflite into library+shell

Also, drive-by change to the test names to keep them matching filenames more closely

* wip

* Update compare_vs_tflite.cpp

* wip

* Fixes

* clang-format

* Fix Makefile

* trigger buildbots
