+++ apps/hannk/util/CMakeLists.txt
@@ -27,3 +27,22 @@ target_include_directories(file_util INTERFACE
 add_library(hannk_log_hdr INTERFACE)
 target_include_directories(hannk_log_hdr INTERFACE
                            $<BUILD_INTERFACE:${hannk_SOURCE_DIR}>)
+
+add_library(model_runner STATIC
+            model_runner.cpp)
+target_include_directories(model_runner PUBLIC
+                           $<BUILD_INTERFACE:${hannk_SOURCE_DIR}>)
+target_compile_definitions(model_runner PRIVATE
+                           -DTFLITE_VERSION_MAJOR=${TFLITE_VERSION_MAJOR}
+                           -DTFLITE_VERSION_MINOR=${TFLITE_VERSION_MINOR}
+                           -DTFLITE_VERSION_PATCH=${TFLITE_VERSION_PATCH})
+target_link_libraries(model_runner PRIVATE
+                      tensorflowlite_c
+                      interpreter
+                      error_util
+                      file_util
+                      hannk_delegate_registrar
+                      tflite_parser
+                      Halide::Tools  # for halide_benchmark.h
+                      Halide::Runtime
+                      tensorflowlite_headers)
